Highlight hpilo kernel module requirement
The ilorest_chif.so library seems to handle a lack of the "hpilo"
kernel module the same way as a lack of iLO devices, returning
"No hpilo devices found" with status 19.
This being converted to a "chif" error causes a misleading
"Chif driver not found" message if the application is executed
on a Linux system having the Chif driver installed, but not having
the "hpilo" module loaded.
At the point the message is printed, the library was already loaded,
indicating it must be correctly present, otherwise it would not have
been able to return an error code.
Improve user guidance by instead setting an error message similar to
the one returned by the library itself, but also pointing out the
need for the hpilo kernel module if not executed on Windows.
Additionally, highlight this requirement in the README.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>